+++ testproject/Assets/Tests/Runtime/SpawnNetworkObjectsDuringSynchronizationTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b4fa2642d341784083366494dd57476
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
